Creating system admin role and applying subsequent changes (#5647)
1) Added new roles to role hierarchy.
2) Refactored system admin role to be able to get all
child roles.
3) User controller modified to only allow deletion or
modification of a user if it has smaller roles than the admin.
4) Refactored employee type to show job title to system admin.
5) Modified user show type to only show roles that
are smaller than the admins role.
6) Refactored partial list template to only show
action icons when the role of the user is smaller
equal to the admins role.
7) Fixed action icons text on the team list page.

+++ src/Opit/Notes/UserBundle/Controller/UserController.php
@@ -21,6 +21,7 @@ use Opit\Notes\UserBundle\Form\ChangePasswordType;
 use Opit\Notes\UserBundle\Entity\User;
 use JMS\SecurityExtraBundle\Annotation\Secure;
 use Opit\Component\Utils\Utils;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 
 /**
  * Description of UserController
@@ -80,12 +81,13 @@ class UserController extends Controller
 
             //create new array for user containing its properties
             $propertyValues[$user->getId()] = array(
-                "username" => $user->getUsername(),
-                "email" => $user->getEmail(),
-                "employeeName" => $employeeName,
-                "isActive" => $user->getIsActive(),
-                "ldapEnabled" => $user->isLdapEnabled(),
-                "roles" => $roles
+                'username' => $user->getUsername(),
+                'email' => $user->getEmail(),
+                'employeeName' => $employeeName,
+                'isActive' => $user->getIsActive(),
+                'ldapEnabled' => $user->isLdapEnabled(),
+                'roles' => $roles,
+                'allowedToEdit' => $this->isSystemAdminAllowedToEdit($user->getId())
             );
         }
 
@@ -144,6 +146,7 @@ class UserController extends Controller
      * @Route("/secured/user/add/{id}", name="OpitNotesUserBundle_user_add", requirements={"id" = "\d+"})
      * @Secure(roles="ROLE_SYSTEM_ADMIN")
      * @Method({"POST"})
+     * @throws AccessDeniedException
      */
     public function addUserAction()
     {
@@ -152,9 +155,11 @@ class UserController extends Controller
         $result = array('response' => 'error');
         $statusCode = 200;
         $errors = array();
-        $isAdmin = $this->get('security.context')->isGranted('ROLE_SYSTEM_ADMIN') ? true : false;
+        $securityContext = $this->container->get('security.context');
+        $isSystemAdmin = $securityContext->isGranted('ROLE_SYSTEM_ADMIN');
         $userService = $this->get('opit.model.user');
-        $id = $isAdmin ? $request->attributes->get('id') : $this->get('security.context')->getToken()->getUser()->getId();
+        $id = $isSystemAdmin ? $request->attributes->get('id') : $this->get('security.context')->getToken()->getUser()->getId();
+        $systemAdminGroups = array();
 
         $user = ($id) ? $this->getUserObject($request->attributes->get('id')) : new User();
 
@@ -163,8 +168,20 @@ class UserController extends Controller
             $user
         );
 
+        // Check if user is system admin
+        if (null !== $user->getId() && $this->isSystemAdmin($securityContext)) {
+            $systemAdminGroups = $this->getSystemAdminGroups($user->getId());
+        }
+
         if ($request->isMethod("POST")) {
             $form->handleRequest($request);
+
+            if(null !== $user->getId() && !$this->isSystemAdminAllowedToEdit($user->getId())) {
+                throw new AccessDeniedException(
+                    'Access denied for user.'
+                );
+            }
+
             // Process form data and create user
             if ($form->isValid()) {
 
@@ -174,13 +191,17 @@ class UserController extends Controller
                     $userService->sendNewPasswordMail($user);
                 }
 
+                foreach ($systemAdminGroups as $systemAdminGroup) {
+                    $user->addGroup(current($systemAdminGroup));
+                }
+
                 // Save the user.
                 $em->persist($user);
                 $em->flush();
 
                 $result['response'] = 'success';
 
-                if ($isAdmin && $request->headers->get('referer') === $this->generateUrl('OpitNotesUserBundle_user_list', array(), true)) {
+                if ($isSystemAdmin && $request->headers->get('referer') === $this->generateUrl('OpitNotesUserBundle_user_list', array(), true)) {
                     return $this->listAction();
                 }
             } else {
@@ -220,6 +241,12 @@ class UserController extends Controller
             foreach ($ids as $id) {
                 // If the logged in user is not equal to deleting user then remove it.
                 if ($loggedInUserId !== (int) $id) {
+                    if(!$this->isSystemAdminAllowedToEdit($id)) {
+                        throw new AccessDeniedException(
+                            'Access denied for user.'
+                        );
+                    }
+
                     $user = $this->getUserObject($id);
                     $em->remove($user);
                 }
@@ -470,4 +497,46 @@ class UserController extends Controller
         return $this->render('OpitNotesUserBundle:User:showUserSummary.html.twig', array('employee' => $user->getEmployee()));
     }
 
+    protected function isSystemAdmin($securityContext)
+    {
+        return $securityContext->isGranted('ROLE_SYSTEM_ADMIN') && !$securityContext->isGranted('ROLE_ADMIN');
+    }
+
+    protected function isSystemAdminAllowedToEdit($userId)
+    {
+        $securityContext = $this->container->get('security.context');
+        $entityManager = $this->getDoctrine()->getManager();
+        $groups = $entityManager->getRepository('OpitNotesUserBundle:Groups');
+        $userRoles = $groups->findUserGroupsArray($userId);
+
+        if (null !== $userId && $this->isSystemAdmin($securityContext) && $securityContext->getToken()->getUser()->getId() != $userId) {
+            foreach ($userRoles as $role) {
+                if (in_array($role['role'], array('ROLE_ADMIN', 'ROLE_GENERAL_MANAGER', 'ROLE_TEAM_MANAGER'))) {
+                    return false;
+                }
+            }
+        }
+
+        return true;
+    }
+
+    protected function getSystemAdminGroups($userId)
+    {
+        $entityManager = $this->getDoctrine()->getManager();
+        $systemAdminGroups = array();
+
+        // Get system admins all assigned groups
+        $systemAdminAssignedGroups = $this->getDoctrine()->getManager()->getRepository('OpitNotesUserBundle:Groups')->findUserGroupsArray($userId);
+        // Assing role system admin
+        $systemAdminGroups[] = $entityManager->getRepository('OpitNotesUserBundle:Groups')->findByRole('ROLE_SYSTEM_ADMIN');
+        // Loop through all roles system admin has got
+        foreach ($systemAdminAssignedGroups as $systemAdminGroup) {
+            // If role is gm or tm add it to the system admin groups array
+            if ('ROLE_TEAM_MANAGER' === $systemAdminGroup['role'] || 'ROLE_GENERAL_MANAGER' === $systemAdminGroup['role']) {
+                $systemAdminGroups[] = $entityManager->getRepository('OpitNotesUserBundle:Groups')->findByRole($systemAdminGroup['role']);
+            }
+        }
+
+        return $systemAdminGroups;
+    }
 }
